backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
flask-socketio>=5.3.6
python-socketio>=5.9.0
eventlet>=0.33.3
websockets>=11.0.3
pandas==1.3.3
numpy==1.21.2
yfinance==0.2.3
requests==2.26.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
nltk>=3.8.1
setuptools>=68.0.0
wheel>=0.41.0
six>=1.16.0
scikit-learn==0.24.2
textblob==0.15.3
python-dateutil==2.8.2